lambda/ingest/requirements.txt
atproto==0.0.55
pydantic>=2.0
pydantic-core>=2.0
fasttext-wheel==0.9.2
numpy<2.0
boto3
